Flytt mapping-funksjoner fra repository til service
